+++ spec/spec_helper.rb
@@ -15,6 +15,7 @@ Lite::Ruby.configure do |config|
   config.object = true
   config.range = true
   config.string = true
+  config.struct = true
   config.time = true
 end